+++ MicroprocessorLib/ReadAbsoluteAddressHigh.cpp
@@ -19,7 +19,8 @@ void CReadAbsoluteAddressHigh::Init(NotMemoryLock bNotMemoryLock, bool bNotVecto
 //////////////////////////////////////////////////////////////////////////
 void CReadAbsoluteAddressHigh::Execute(CW65C816* pcCPU)
 {
-	pcCPU->GetState()->SetAddressHigh(pcCPU->GetState()->GetData());
+	//Is this the right Data?
+	pcCPU->GetState()->SetAddressHigh((uint8)pcCPU->GetState()->GetData());
 }
