+++ MicroprocessorLib/ReadAbsoluteAddressLow.cpp
@@ -19,7 +19,8 @@ void CReadAbsoluteAddressLow::Init(NotMemoryLock bNotMemoryLock, bool bNotVector
 //////////////////////////////////////////////////////////////////////////
 void CReadAbsoluteAddressLow::Execute(CW65C816* pcCPU)
 {
-	pcCPU->GetState()->SetAddressLow(pcCPU->GetState()->GetData());
+	//Is this the right Data?
+	pcCPU->GetState()->SetAddressLow((uint8)pcCPU->GetState()->GetData());
 }
